issue-34331 修改统一订单的接口，跟钱有关的字段，都改成以分为单位了（类型为Long）

+++ ehcore/src/main/java/com/everhomes/asset/DefaultAssetVendorHandler.java
@@ -113,7 +113,7 @@ public class DefaultAssetVendorHandler extends AssetVendorHandler{
         CreatePurchaseOrderCommand preOrderCommand = new CreatePurchaseOrderCommand();
         
         BigDecimal totalAmountCents = calculateBillOrderAmount(cmd);
-        preOrderCommand.setAmount(totalAmountCents);
+        preOrderCommand.setAmount(totalAmountCents.longValue());
         
         String accountCode = generateAccountCode(cmd.getNamespaceId());
         preOrderCommand.setAccountCode(accountCode);
